Fix signed/unsigned mismatch.
SVN-Revision: 690

+++ libarchive/test/test_read_pax_truncated.c
@@ -30,7 +30,7 @@ DEFINE_TEST(test_read_pax_truncated)
 	struct archive_entry *ae;
 	struct archive *a;
 	size_t used, i, buff_size = 1000000;
-	ssize_t filedata_size = 100000;
+	size_t filedata_size = 100000;
 	char *buff = malloc(buff_size);
 	char *buff2 = malloc(buff_size);
 	char *filedata = malloc(filedata_size);
@@ -56,7 +56,8 @@ DEFINE_TEST(test_read_pax_truncated)
 	archive_entry_set_size(ae, filedata_size);
 	assertA(0 == archive_write_header(a, ae));
 	archive_entry_free(ae);
-	assertA(filedata_size == archive_write_data(a, filedata, filedata_size));
+	assertA((ssize_t)filedata_size
+	    == archive_write_data(a, filedata, filedata_size));
 
 	/* Close out the archive. */
 	assertA(0 == archive_write_close(a));
